pairing check
