+++ packages/ultrahonk/tests/plain.rs
@@ -1,6 +1,7 @@
 use ark_bn254::Bn254;
 use ark_ec::AffineRepr;
 use ark_ec::{pairing::Pairing, CurveGroup};
+use ark_ff::One;
 use ark_ff::{BigInt, Field};
 use ark_serialize::CanonicalDeserialize;
 use eyre::{anyhow, Result};
@@ -25,7 +26,6 @@ use ultrahonk::{
     prelude::{HonkProof, UltraHonk},
     types::ScalarField,
 };
-use ark_ff::One;
 
 pub struct ArkKeccak256;
 
@@ -93,12 +93,18 @@ impl G1ArithmeticBackend for ArkHonkCurve {
 
     /// Check the pairing identity e(a_1, b_1) == e(a_2, b_2)
     fn ec_pairing_check(
-        a_1: G1Affine,
-        b_1: G2Affine,
-        a_2: G1Affine,
-        b_2: G2Affine,
+        p0: G1Affine,
+        p1: G1Affine,
+        g2_x: G2Affine,
+        g2_gen: G2Affine,
     ) -> Result<bool, G1ArithmeticError> {
-        Ok(<Bn254 as Pairing>::multi_pairing([a_1, a_2], [b_1, b_2]).0
+        tracing::trace!("Pairing check");
+        let p = [g2_gen, g2_x];
+        let g1_prepared = [
+            <Bn254 as Pairing>::G1Prepared::from(p0),
+            <Bn254 as Pairing>::G1Prepared::from(p1),
+        ];
+        Ok(<Bn254 as Pairing>::multi_pairing(g1_prepared, p).0
             == <Bn254 as Pairing>::TargetField::one())
     }
 
@@ -154,7 +160,7 @@ impl HonkCurve for ArkHonkCurve {
     }
 }
 
-fn plain_test<H: HashBackend>(proof_file: &str, vk_file: &str, public_inputs_file: &str) {
+fn plain_test(proof_file: &str, vk_file: &str, public_inputs_file: &str) {
     const CRS_PATH_G2: &str = "./src/crs/bn254_g2.dat";
 
     // parse proof file
@@ -173,9 +179,14 @@ fn plain_test<H: HashBackend>(proof_file: &str, vk_file: &str, public_inputs_fil
     let vk = VerifyingKeyBarretenberg::from_buffer(&vk_u8).unwrap();
     let vk = VerifyingKey::from_barrettenberg_and_crs(vk, verifier_crs);
 
-    let is_valid =
-        UltraHonk::<ArkHonkCurve, H>::verify(proof, &public_inputs, &vk, ZeroKnowledge::No)
-            .unwrap();
+    let is_valid = UltraHonk::<ArkHonkCurve, ArkKeccak256>::verify(
+        proof,
+        &public_inputs,
+        &vk,
+        ZeroKnowledge::No,
+    )
+    .unwrap();
+
     assert!(is_valid);
 }
 
@@ -197,6 +208,6 @@ fn test_iterating_test_vectors() {
         let vk_file = format!("{}/kat/vk", path.display());
         let public_inputs_file = format!("{}/kat/public_inputs", path.display());
 
-        plain_test::<ArkKeccak256>(&proof_file, &vk_file, &public_inputs_file);
+        plain_test(&proof_file, &vk_file, &public_inputs_file);
     }
 }
